# https://github.com/commaai/openpilot/issues/32536 Speedup test_car_interfaces.py