https://github.com/commaai/openpilot/issues/32536 Speedup test_car_interfaces.py

**Description**
Increased the slowest test from ~1-1.5s seconds to a max of ~0.6s and total test execution time from ~16.5s to ~9.8s on M1 Pro (10 cores)

This was achieved by applying 3 optimizations:

1. Build `hypothesis` strategies once instead of rebuilding them on each test case run: total execution time down from ~16.5s to ~12.2s
2. Precompute `K1FD` parameters for `CarStateBase.__init__()`: total time execution time down from ~12.2s to ~9.8s
3. Ford-specific optimization: cache `CANDefine` objects created during `CarState` construction, which increased the per-car execution time from 0.8s (given all the above optimizations) to ~0.54s


**Next Steps**
1. I plan to root cause the `hypothesis` slowdown, which was repeatedly mentioned in the issue discussion.
2. I plan to explore whether the 3rd optimization improves performance for other car models.