@adeebshihadeh 2025-05-14
So this takes the slowest test on my workstation from 1.33s to 0.86s. Definitely faster, but still pretty far off from the target in the bounty.