@blackoutnet 2025-05-14
@adeebshihadeh, ok. One question, though: Did you apply the opendbc update? It should be closer to 0.6s or 0.5s.
Could you, please, post the pytest log?